update dtype handling

+++ src/ShAReD_Net/model/activation/base.py
@@ -3,8 +3,8 @@ import time
 import tensorflow as tf
 
 class DiscretSigmoid(tf.keras.layers.Layer):
-    def __init__(self, name = "DiscretSigmoid",dtype=tf.float32, **kwargs):
-        super().__init__(name = name, dtype = dtype, **kwargs)
+    def __init__(self, name = "DiscretSigmoid", **kwargs):
+        super().__init__(name = name, **kwargs)
     
     @tf.function(experimental_relax_shapes=True)
     def call(self, inputs, training=None):
